fix: comprehensive review fixes across Rust, TypeScript, tests, and config
Rust crates:
- Fix f64_to_i64 boundary check (>= instead of > for i64::MAX as f64)
- Add f64_to_i128 with integrality validation
- WASM: implement BrowserHostSystem using js_sys::Date::now() for now_*
- WASM: use f64_to_i64/f64_to_i128 helpers in Instant/ZonedDateTime
- WASM: proper js_sys::Error objects instead of plain string errors
- WASM: add #[wasm_bindgen(inspectable)] to all structs
- NAPI: add toString rounding options to Instant and ZonedDateTime
- NAPI: remove double-reference on provider in now.rs
- Fix Duration::sign() return type from i8 to i32
- Add precision-loss doc comments to WASM Duration getters
- Consistent overflow handling in PlainYearMonth add/subtract

TypeScript lib:
- Default undefined sub-second fields to 0 in toNapiPlainTime
- Duration constructor: named params instead of arguments[0..9]
- Move DURATION_FIELDS_CHECK to module-level constant
- ZonedDateTime constructor: only catch RangeError, rethrow others
- Add requireBranding to PlainDate/Duration/PlainTime add/subtract
- Add requireBranding to ZonedDateTime.with()
- Null guards for _origFormatGetter in 6 toLocaleString methods
- Fix comment: "7-day steps" → "6-day steps" in timezone.ts
- Extract _estimateIsoYear helper in PlainMonthDay.from() (DRY)
- Unify validateMonthCodeSyntax to return parsed components
- Fix PlainDateTime.from() validateOptions consistency
- Validate calendar identifier format in getCalendarId

Tests:
- Test262 runner: verify error type matches for negative tests
- Test262 runner: create failures file early for crash resilience
- Add error type assertions to bare .toThrow() calls
- Replace hardcoded year 2024 with dynamic getFullYear() - 1
- Expand cjs.spec.ts to test all major exports
- Add tests for with(), round(), toString options, toLocaleString
- Add Duration.total()/round() with relativeTo tests

Config/build:
- Package.json: fix ./native export for ESM consumers
- tsup: add native binding path to external list
- Add tsconfig.test.json for test file type-checking
- ESLint: lint test files with relaxed rules
- CI: add npm audit, guard test262-failures.txt, debug build for lint
- Cargo.toml: add codegen-units = 1 to release profile

+++ crates/temporal-napi/src/duration.rs
@@ -109,19 +109,21 @@ impl Duration {
         self.inner.milliseconds() as f64
     }
 
+    /// Note: returns f64 per TC39 Temporal spec. Precision loss possible for values > 2^53.
     #[napi(getter)]
     pub fn microseconds(&self) -> f64 {
         self.inner.microseconds() as f64
     }
 
+    /// Note: returns f64 per TC39 Temporal spec. Precision loss possible for values > 2^53.
     #[napi(getter)]
     pub fn nanoseconds(&self) -> f64 {
         self.inner.nanoseconds() as f64
     }
 
     #[napi(getter)]
-    pub fn sign(&self) -> i8 {
-        self.inner.sign() as i8
+    pub fn sign(&self) -> i32 {
+        self.inner.sign() as i32
     }
 
     #[napi(getter)]
